// src/domains/posts/controller/posts.controller.interface.ts
import { NextFunction, Request, Response } from 'express';
import { BaseController } from '../../../common/base.controller';

export interface IPostsController extends BaseController {
  create: (req: Request, res: Response, next: NextFunction) => Promise<void>;
  find: (req: Request, res: Response, next: NextFunction) => Promise<void>;
  update: (req: Request, res: Response, next: NextFunction) => Promise<void>;
  delete: (req: Request, res: Response, next: NextFunction) => Promise<void>;
}